+++ de.metas.fresh/de.metas.fresh.base/src/main/java/de/metas/fresh/picking/PackingMaterialKey.java
@@ -81,6 +81,7 @@ public class PackingMaterialKey extends TerminalKey
 		displayName = piItemProductBL
 				.buildDisplayName()
 				.setM_HU_PI_Item_Product(itemProduct)
+				.setShowAnyProductIndicator(true)
 				.buildItemProductDisplayName();
 		value = new KeyNamePair(piItemProductId, displayName);
 	}